+++ tests-python/test_smoke.py
@@ -2,4 +2,4 @@ import lapinbeam
 
 
 def test_version():
-    assert lapinbeam.__version__ == "1.0.0"
+    assert lapinbeam.__version__ == "1.0.1"